+++ src/main/java/com/imsweb/algorithms/Algorithms.java
@@ -27,6 +27,7 @@ import com.imsweb.algorithms.ruralurban.UrbanContinuumAlgorithm;
 import com.imsweb.algorithms.ruralurban.UricAlgorithm;
 import com.imsweb.algorithms.seersiterecode.SeerSiteRecodeAlgorithm;
 import com.imsweb.algorithms.survival.SurvivalTimeAlgorithm;
+import com.imsweb.algorithms.tractestcongressdist.TractEstCongressDistAlgorithm;
 import com.imsweb.algorithms.yostacspoverty.YostAcsPovertyAlgorithm;
 
 import static com.imsweb.algorithms.AlgorithmField.DATA_LEVEL_PATIENT;
@@ -61,6 +62,7 @@ public class Algorithms {
     public static final String ALG_PRCDA_UIHO = "prcda-uiho";
     public static final String ALG_ACS_LINKAGE = "acs-linkage";
     public static final String ALG_EPHT_SUBCOUNTY = "epht-subcounty";
+    public static final String ALG_TRACT_EST_CONGRESS_DIST = "tract-est-congress-dist";
 
     // special properties
     public static final String FIELD_TUMORS = "tumors";
@@ -177,6 +179,7 @@ public class Algorithms {
     public static final String FIELD_ACS_POV1418_HISPANIC = "acsPctPov1418Hispanic";
     public static final String FIELD_EPHT_2010_GEOID_5K = "epht2010GeoId5k";
     public static final String FIELD_EPHT_2010_GEOID_20K = "epht2010GeoId20k";
+    public static final String FIELD_TRACT_EST_CONGRESS_DIST = "tractEstCongressDist";
 
     // options
     public static final String PARAM_NHIA_OPTION = "nhiaOption";
@@ -311,7 +314,9 @@ public class Algorithms {
             addField(AlgorithmField.of(FIELD_ACS_POV1418_HISPANIC, null, 6, "ACS Pct Poverty (2014-2018, Hispanic)", "ACS Pov 1418 Hispanic", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_EPHT_2010_GEOID_5K, null, 11, "EPHT 2010 GEO ID 5K", "EPHT 5K", DATA_LEVEL_TUMOR));
             addField(AlgorithmField.of(FIELD_EPHT_2010_GEOID_20K, null, 11, "EPHT 2010 GEO ID 20K", "EPHT 20K", DATA_LEVEL_TUMOR));
-            
+
+            addField(AlgorithmField.of(FIELD_TRACT_EST_CONGRESS_DIST, null, 2, "Tract-Estimated Congressional District", "Tract-Est Congress Dist", DATA_LEVEL_TUMOR));
+
             // algorithms
             addAlgorithm(new NhiaAlgorithm());
             addAlgorithm(new NapiiaAlgorithm());
@@ -330,6 +335,7 @@ public class Algorithms {
             addAlgorithm(new PrcdaUihoAlgorithm());
             addAlgorithm(new YostAcsPovertyAlgorithm());
             addAlgorithm(new EphtSubCountyAlgorithm());
+            addAlgorithm(new TractEstCongressDistAlgorithm());
         }
         finally {
             _LOCK.writeLock().unlock();